separate test into sets of unit, integration and components
