+++ order-app/src/test/java/unit/com/perficient/orderapp/domain/CartTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
+import java.util.UUID;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -20,6 +21,24 @@ public class CartTest {
         var product2 = ProductItemMother.product2
                 .price(BigDecimal.valueOf(3.6)).build();
         var cart = new Cart();
+        // WHEN
+        cart.addProduct(product1);
+        cart.addProduct(product2);
+        var total = cart.getTotalPrice();
+        // THEN
+        assertEquals(BigDecimal.valueOf(8.1), total);
+    }
+
+    @Test
+    @DisplayName("Empty Cart retrieve from db  \n" +
+            "should add products success")
+    void add_product_empty_cart_success() {
+        // GIVEN
+        var product1 = ProductItemMother.product1
+                .price(BigDecimal.valueOf(4.5)).build();
+        var product2 = ProductItemMother.product2
+                .price(BigDecimal.valueOf(3.6)).build();
+        var cart = new Cart(UUID.randomUUID(), null, null);
         // WHEN
         cart.addProduct(product1);
         cart.addProduct(product2);
